Update and rename install.sh to exec.sh

+++ exec.sh
@@ -2,13 +2,19 @@
 
 set -e
 
+REPO_URL="https://github.com/KiamMota/kimasnvim.git"
 NVIM_DIR="$HOME/.config/nvim"
 
-echo ">> removing old nvim config"
-rm -rf "$NVIM_DIR"
-
-echo ">> cloning KIMASNVIM"
-git clone https://github.com/KiamMota/kimasnvim.git "$NVIM_DIR"
+if [ -d "$NVIM_DIR/.git" ]; then
+  echo ">> KIMASNVIM already exists"
+  echo ">> pulling latest changes"
+  cd "$NVIM_DIR"
+  git pull --ff-only
+else
+  echo ">> installing KIMASNVIM"
+  rm -rf "$NVIM_DIR"
+  git clone "$REPO_URL" "$NVIM_DIR"
+fi
 
 clear
 
@@ -23,5 +29,4 @@ wellcome to
 
 EOF
 
-echo ">> KIMASNVIM installed"
-echo ">> open nvim and let it breathe"
+echo ">> KIMASNVIM ready"
